Its time to make money task 9

+++ controller/expence.js
@@ -7,6 +7,7 @@ exports.addExpense = async (req,res)=>{
         const expenseamount = req.body.expenseamount;
         const description = req.body.description;
         const category = req.body.category;
+        const userId = req.user.id
         
         if(expenseamount == undefined || expenseamount.length === 0 ){
             return res.status(400).json({success: false, message: 'Parameters missing'})
@@ -15,6 +16,7 @@ exports.addExpense = async (req,res)=>{
             expenseamount:expenseamount,
             category:category,
             description:description,
+            userId: userId,
         })
         return res.status(201).json({expence:expence,success:true});
     }
@@ -24,8 +26,8 @@ exports.addExpense = async (req,res)=>{
 }
 exports.getExpenses = async(req,res)=>{
     try {
-        // const expenses = await Expense.findAll({where:{userId: req.body.id}})
-        const expenses = await Expense.findAll()
+        const expenses = await Expense.findAll({where:{userId: req.user.id}})
+        //const expenses = await Expense.findAll()
         return res.status(200).json({expences:expenses,success:true});
     } catch (error) {
         return res.status(200).json({error:error,success:false})
@@ -40,8 +42,8 @@ exports.deleteExpense = async (req,res)=>{
             return res.status(400).json({success: false, })
         }
         console.log(expenseid);
-        // await Expense.destroy({where:{id:expenseid,userId:req.body.id}})
-        await Expense.destroy({where:{id:expenseid}})
+        await Expense.destroy({where:{id:expenseid,userId:req.user.id}})
+        //await Expense.destroy({where:{id:expenseid}})
         return res.status(200).json({success:true,message:"Deleted Successfully"})
     } catch (error) {
         return res.status(500).json({ success: true, message: "Failed"})
